changed inputChecker to parseArgs and moved it to a new file
